Curso de postman/postman-course/requirements.txt
backcall==0.1.0
decorator==4.4.1
Django==2.2.7
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7